refactor(task): renamed use case file
Renamed use case file from "delete-task-by-id" to "delete-task-by-id-use-case"
