+++ src/domain/use-cases/delete-task-by-id-use-case/delete-task-by-id-use-case.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import { Task } from '../../entities/task/task';
-import { DeleteTaskByIdUseCase } from './delete-task-by-id';
+import { DeleteTaskByIdUseCase } from './delete-task-by-id-use-case';
 import {
   IDeleteTaskByIdRepository,
   IFetchTaskByIdRepository,
