+++ src/domain/use-cases/index.ts
@@ -1,2 +1,2 @@
-export * from './delete-task-by-id/delete-task-by-id';
+export * from './delete-task-by-id-use-case/delete-task-by-id-use-case';
 export * from './update-task-by-id-use-case/update-task-by-id-use-case';
